+++ flow-typed/types/provider-options.js
@@ -10,5 +10,6 @@ declare type ProviderOptionsObject = {
   env?: ProviderEnvConfigObject,
   networkRetryParameters?: ProviderNetworkRetryParameters,
   filterOptions?: ProviderFilterOptionsObject,
-  ignoreServerConfig?: boolean
+  ignoreServerConfig?: boolean,
+  loadThumbnailWithKs?: boolean
 };